+++ packages/utilities/postbuild.js
@@ -1,0 +1,3 @@
+const postbuild = require('../../scripts/postbuild');
+
+postbuild(__dirname)