+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
 ### Added
 - `CHANGELOG.md` file
 - `root`: `package.json`, `.editorconfig`, `.gitattributes`, `.gitignore` initial files
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `root`: `@distributedlab/utilities` package particularly moved from the [old repo] `@distributedlab/utils` package
 - `root`: `@distributedlab/jac` package moved from the [old repo] `@distributedlab/json-api-client` package
 - `root`: GitHub Actions
+- `all`: Build to CommonJS and ES modules
+- `all`: `tsc-alias` package to use aliases in TypeScript
+- `root`: `yarn rsc` Release Sanity Check script
+- `root`: Rollup and configuration file to build packages for CDN
 
 ### Changed
 - `root`: Updated `README.md`